mod_fastcgi.c: write error log only if seek position is greater than the file lenght
As BAP works, it downloads the .zat file and polls periodically if the file
grew. If the connection is idle, BAP always queries the end of the file, which
previously always produced an error log message. Now this log message is
written only if the query is reading the file at EOF+n where n>1.